feature/watchlist: adjusting test according to the new props in analysis screen
